Moved all printing logic to PrettyPrinter. Additionally, removed this logic from main.
